# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Envelope:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
    type: object
...
  watchlist.AddStockRequest:
    properties:
      code:
        example: 600519.SH
        type: string
      market:
        description: Market 留空时由代码自动推断，只有代码本身不带后缀时才需要显式给出。
        example: CN
        type: string
      note:
        example: 业绩拐点，盯一下
        type: string
    required:
      - code
    type: object
  watchlist.DeletedView:
    properties:
      deleted:
        example: true
        type: boolean
    type: object
  watchlist.GroupDetailView:
    properties:
      group:
        $ref: '#/definitions/watchlist.GroupView'
      items:
        items:
          $ref: '#/definitions/watchlist.ItemView'
        type: array
    type: object
  watchlist.GroupNameRequest:
    properties:
      name:
        example: 科技股
        type: string
    required:
      - name
    type: object
  watchlist.GroupView:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      itemCount:
        type: integer
      name:
        example: 科技股
        type: string
      updatedAt:
        type: string
    type: object
  watchlist.ItemView:
    properties:
      changePct:
        type: string
      createdAt:
        type: string
      gainPctSinceAdded:
        type: string
      market:
        example: CN
        type: string
      note:
        type: string
      price:
        description: |-
          行情缺失时以下字段为 null，前端显示「暂无行情」，而不是 0。

          类型是 *string 而不是 *float64：指针保留「有没有这个数」的语义，
          字符串保留精度。两者都不能省——用 float64 则 12.34 到前端可能变成
          12.339999999999999，用非指针则「停牌没有行情」和「价格为 0」再也分不开。
        type: string
      referenceDate:
        type: string
      referencePrice:
        type: string
      sortOrder:
        type: integer
      symbol:
        example: "600519"
        type: string
      tradeDate:
        type: string
    type: object
  watchlist.MoveStockRequest:
    properties:
      code:
        example: 600519.SH
        type: string
      fromGroupId:
        type: integer
      market:
        example: CN
        type: string
      toGroupId:
        type: integer
    required:
      - code
      - fromGroupId
      - toGroupId
    type: object
  watchlist.MovedView:
    properties:
      moved:
        example: true
        type: boolean
    type: object
  watchlist.RemovedView:
    properties:
      removed:
        example: true
        type: boolean
    type: object
  watchlist.ReorderRequest:
    properties:
      codes:
        example:
          - 600519.SH
          - 000001.SZ
        items:
          type: string
        type: array
    required:
      - codes
    type: object
  watchlist.ReorderedView:
    properties:
      reordered:
        example: true
        type: boolean
    type: object
  watchlist.UpdateNoteRequest:
    properties:
      note:
        example: 业绩拐点，盯一下
        type: string
    type: object
  watchlist.UpdatedView:
    properties:
      updated:
        example: true
        type: boolean
    type: object
info:
  contact: { }
  description: |-
    多智能体股票分析平台 HTTP 接口。

    所有响应共用同一个信封 {"code":0,"message":"ok","data":{...}}，
    code 为 0 表示成功，非 0 时 data 恒为 null，具体取值见 response 包。

    健康检查 GET /healthz 不在 /api/v1 之下，因此未收录于本文档。

    免责声明：本平台仅用于学习与研究，不构成投资建议。分析结论由大模型生成，
    可能包含事实性错误。过往表现不代表未来收益，投资有风险，可能损失本金。
  title: Trading Agents API
  version: "1.0"
paths:
...
  /watchlist/groups:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/watchlist.GroupView'
                    type: array
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 分组列表
      tags:
        - 自选股
    post:
      consumes:
        - application/json
      parameters:
        - description: 分组名
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.GroupNameRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.GroupView'
                type: object
        "400":
          description: 分组名为空、超过 16 个字符或含控制字符
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 已存在同名分组
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 新建分组
      tags:
        - 自选股
  /watchlist/groups/{id}:
    delete:
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.DeletedView'
                type: object
        "400":
          description: 分组 ID 非法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 删除分组
      tags:
        - 自选股
    put:
      consumes:
        - application/json
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 新分组名
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.GroupNameRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.GroupView'
                type: object
        "400":
          description: 分组 ID 非法，或分组名为空、超长、含控制字符
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 已存在同名分组，或分组已被并发修改
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 重命名分组
      tags:
        - 自选股
  /watchlist/groups/{id}/items:
    get:
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.GroupDetailView'
                type: object
        "400":
          description: 分组 ID 非法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 分组明细
      tags:
        - 自选股
    post:
      consumes:
        - application/json
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 标的与备注
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.AddStockRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.ItemView'
                type: object
        "400":
          description: 分组 ID 或标的代码非法，或备注超过 100 个字符
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 该标的已在分组内
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 分组已达 200 只上限
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 添加自选股
      tags:
        - 自选股
  /watchlist/groups/{id}/items/{code}:
    delete:
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 标的代码，形如 600519.SH；点号无需转义，直接放进路径即可
          in: path
          name: code
          required: true
          type: string
        - description: 市场（CN/HK/US），留空时由代码自动推断
          in: query
          name: market
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.RemovedView'
                type: object
        "400":
          description: 分组 ID 或标的代码非法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在、不属于当前用户，或该标的不在分组内
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 分组已被并发修改，请重新加载后重试
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 移除自选股
      tags:
        - 自选股
  /watchlist/groups/{id}/items/{code}/note:
    put:
      consumes:
        - application/json
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 标的代码，形如 600519.SH；点号无需转义，直接放进路径即可
          in: path
          name: code
          required: true
          type: string
        - description: 市场（CN/HK/US），留空时由代码自动推断
          in: query
          name: market
          type: string
        - description: 备注，传空串表示清空
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.UpdateNoteRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.UpdatedView'
                type: object
        "400":
          description: 分组 ID 或标的代码非法，或备注超过 100 个字符
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在、不属于当前用户，或该标的不在分组内
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 分组已被并发修改，请重新加载后重试
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 修改自选股备注
      tags:
        - 自选股
  /watchlist/groups/{id}/items/reorder:
    post:
      consumes:
        - application/json
      parameters:
        - description: 分组 ID
          in: path
          name: id
          required: true
          type: integer
        - description: 期望排在最前面的标的代码，按新顺序给出
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.ReorderRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.ReorderedView'
                type: object
        "400":
          description: 分组 ID 非法，或排序列表为空、含非法代码、含重复代码、含组内不存在的标的
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 分组不存在或不属于当前用户
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 分组已被并发修改，请重新加载后重试
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 重排自选股顺序
      tags:
        - 自选股
  /watchlist/items/move:
    post:
      consumes:
        - application/json
      parameters:
        - description: 源分组、目标分组与标的
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/watchlist.MoveStockRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/watchlist.MovedView'
                type: object
        "400":
          description: 标的代码非法，或源分组与目标分组相同
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 源或目标分组不存在、不属于当前用户，或该标的不在源分组内
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 目标分组中已存在该标的
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 目标分组已达 200 只上限
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 跨分组移动自选股
      tags:
        - 自选股
securityDefinitions:
  BearerAuth:
    description: 填入 "Bearer {accessToken}"，令牌由 POST /auth/login 获取。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...